fix: skip override on exports
