+++ test/module/esm-import.js
@@ -1,48 +1,62 @@
 import t from 'tap'
 import fastify from 'fastify'
-import esmImportApp from './esm-import/app.js'
+import esmImportAppDefault from './esm-import/app-default.js'
+import esmImportAppNamed from './esm-import/app-named.js'
+import esmImportAppStarDefault from './esm-import/app-star-default.js'
+import esmImportAppStarNamed from './esm-import/app-star-named.js'
 
-t.plan(13)
+const { test } = t
 
-const app = fastify()
+test('default', async () => {
+  const app = fastify()
 
-app.register(esmImportApp)
+  app.register(esmImportAppDefault)
+  
+  await app.ready() 
+  
+  const res = await app.inject('/default') 
+  t.equal(res.statusCode, 200)
+  t.deepEqual(res.json(), { script: 'default' })
+})
 
-app.ready(function (err) {
-  t.error(err)
 
-  app.inject({
-    url: '/named'
-  }, function (err, res) {
-    t.error(err)
+test('named', async () => {
+  const app = fastify()
 
-    t.equal(res.statusCode, 200)
-    t.deepEqual(JSON.parse(res.payload), { script: 'named' })
-  })
+  app.register(esmImportAppNamed)
+  
+  await app.ready()
+  
+  const res = await app.inject('/named')
 
-  app.inject({
-    url: '/default'
-  }, function (err, res) {
-    t.error(err)
+  t.equal(res.statusCode, 200)
+  t.deepEqual(res.json(), { script: 'named' })
+})
 
-    t.equal(res.statusCode, 200)
-    t.deepEqual(JSON.parse(res.payload), { script: 'default' })
-  })
 
-  app.inject({
-    url: '/star-default'
-  }, function (err, res) {
-    t.error(err)
+test('star-default', async () => {
+  const app = fastify()
 
-    t.equal(res.statusCode, 200)
-    t.deepEqual(JSON.parse(res.payload), { script: 'star-default' })
-  })
-  app.inject({
-    url: '/star-named'
-  }, function (err, res) {
-    t.error(err)
+  app.register(esmImportAppStarDefault)
+  
+  await app.ready()
+  
+  const res = await app.inject('/star-default')
 
-    t.equal(res.statusCode, 200)
-    t.deepEqual(JSON.parse(res.payload), { script: 'star-named' })
-  })
+  t.equal(res.statusCode, 200)
+  t.deepEqual(res.json(), { script: 'star-default' })
+})
+
+
+test('star-named', async () => {
+  const app = fastify()
+
+  app.register(esmImportAppStarNamed)
+  
+  await app.ready()
+
+  const res = await app.inject('/star-named')
+
+  t.equal(res.statusCode, 200)
+  t.deepEqual(res.json(), { script: 'star-named' })
 })
