+++ test/module/esm-import/app-star-default.js
@@ -1,0 +1,16 @@
+import path from 'path'
+import { fileURLToPath } from 'url'
+import { dirname } from 'path'
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+import * as fastifyAutoloadStar from '../../../index.js'
+
+export default async function (fastify, opts) {
+
+  fastify.register(fastifyAutoloadStar.default, {
+    dir: path.join(__dirname, 'plugins', 'star-default'),
+    options: { foo: 'bar' }
+  })
+}
